+++ test/samples/arrow-functions.js
@@ -171,11 +171,12 @@ module.exports = [
 			() => console.log( 'not printed' );`,
 
 		output: `
-			(function() { return console.log( 'not printed' ); });`
+			!function() { return console.log( 'not printed' ); };`
 	},
 
 	{
-		description: 'handles standalone arrow function expression statement within a function',
+		description:
+			'handles standalone arrow function expression statement within a function',
 
 		input: `
 			function no_op () {
@@ -184,7 +185,25 @@ module.exports = [
 
 		output: `
 			function no_op () {
-				(function() { return console.log( 'not printed' ); });
+				!function() { return console.log( 'not printed' ); };
 			}`
+	},
+
+	{
+		description:
+			'are transformed even if disabled if they have a transpiled spread parameter',
+
+		options: { transforms: { arrow: false, spreadRest: true } },
+
+		input: `
+				(...args) => console.log( args );`,
+
+		output: `
+				!function() {
+					var args = [], len = arguments.length;
+					while ( len-- ) args[ len ] = arguments[ len ];
+
+					return console.log( args );
+				};`
 	}
 ];