+++ test/samples/classes.js
@@ -40,7 +40,8 @@ module.exports = [
 	},
 
 	{
-		description: 'transpiles a class declaration without a constructor function',
+		description:
+			'transpiles a class declaration without a constructor function',
 
 		input: `
 			class Foo {
@@ -343,7 +344,8 @@ module.exports = [
 	},
 
 	{
-		description: 'transpiles an anonymous class expression with a non-constructor method',
+		description:
+			'transpiles an anonymous class expression with a non-constructor method',
 
 		input: `
 			var Foo = class {
@@ -361,6 +363,29 @@ module.exports = [
 				};
 
 				return Foo;
+			}());`
+	},
+
+	{
+		description:
+			'transpiles an anonymous class expression that is assigned to a property',
+
+		input: `
+			const q = {};
+
+			q.a = class {
+				c () {}
+			};`,
+
+		output: `
+			var q = {};
+
+			q.a = (function () {
+				function a () {}
+
+				a.prototype.c = function c () {};
+
+				return a;
 			}());`
 	},
 
@@ -656,7 +681,8 @@ module.exports = [
 	},
 
 	{
-		description: 'verify deindent() does not corrupt string literals in class methods (#159)',
+		description:
+			'verify deindent() does not corrupt string literals in class methods (#159)',
 
 		input: `
 			class Foo {
@@ -811,7 +837,8 @@ module.exports = [
 	},
 
 	{
-		description: 'uses correct indentation for inserted statements in constructor (#39)',
+		description:
+			'uses correct indentation for inserted statements in constructor (#39)',
 
 		input: `
 			class Foo {
@@ -854,7 +881,8 @@ module.exports = [
 	},
 
 	{
-		description: 'uses correct indentation for inserted statements in subclass constructor (#39)',
+		description:
+			'uses correct indentation for inserted statements in subclass constructor (#39)',
 
 		input: `
 			class Foo extends Bar {
@@ -1038,7 +1066,7 @@ module.exports = [
 
 			Foo.prototype.catch = function catch$1 () {
 				// code goes here
-			};`,
+			};`
 	},
 
 	{
@@ -1113,7 +1141,8 @@ module.exports = [
 	},
 
 	{
-		description: 'static methods with computed names with varied spacing (#139)',
+		description:
+			'static methods with computed names with varied spacing (#139)',
 
 		input: `
 			class B {
@@ -1163,7 +1192,8 @@ module.exports = [
 	},
 
 	{
-		description: 'static methods with numeric or string names with varied spacing (#139)',
+		description:
+			'static methods with numeric or string names with varied spacing (#139)',
 
 		input: `
 			class D {
@@ -1196,7 +1226,7 @@ module.exports = [
 			X.prototype.foo = function foo$1 () { return foo };
 			X.prototype.bar = function bar () {};
 		`
-	},
+	}
 
 	// TODO more tests. e.g. getters and setters.
 	// 'super.*' is not allowed before super()