+++ test/samples/computed-properties.js
@@ -35,10 +35,41 @@ module.exports = [
 			};`,
 
 		output: `
+			var obj = {};
+			obj[a] = 1;
+			obj.b = 2;`
+	},
+
+	{
+		description: 'creates a computed property at start of literal with method afterwards',
+
+		input: `
 			var obj = {
-				b: 2
-			};
-			obj[a] = 1;`
+				[a]: 1,
+				b() {}
+			};`,
+
+		output: `
+			var obj = {};
+			obj[a] = 1;
+			obj.b = function b() {};`
+	},
+
+	{
+		description: 'creates a computed property at start of literal with generator method afterwards when transpiling methods is disabled',
+
+		options: { transforms: { conciseMethodProperty: false, generator: false } },
+
+		input: `
+			var obj = {
+				[a]: 1,
+				*b() {}
+			};`,
+
+		output: `
+			var obj = {};
+			obj[a] = 1;
+			obj.b = function* () {};`
 	},
 
 	{
@@ -69,10 +100,10 @@ module.exports = [
 
 		output: `
 			var obj = {
-				a: 1,
-				c: 3
+				a: 1
 			};
-			obj[b] = 2;`
+			obj[b] = 2;
+			obj.c = 3;`
 	},
 
 	{
@@ -89,13 +120,12 @@ module.exports = [
 			};`,
 
 		output: `
-			var obj = {
-				b: 2,
-				e: 5
-			};
+			var obj = {};
 			obj[a] = 1;
+			obj.b = 2;
 			obj[c] = 3;
 			obj[d] = 4;
+			obj.e = 5;
 			obj[f] = 6;`
 	},
 
@@ -106,8 +136,9 @@ module.exports = [
 			var a = 'foo', obj = { [a]: 'bar', x: 42 }, bar = obj.foo;`,
 
 		output: `
-			var a = 'foo', obj = ( _obj = { x: 42 }, _obj[a] = 'bar', _obj ), bar = obj.foo;
-			var _obj;`
+			var _obj;
+
+			var a = 'foo', obj = ( _obj = {}, _obj[a] = 'bar', _obj.x = 42, _obj ), bar = obj.foo;`
 	},
 
 	{
@@ -138,8 +169,9 @@ module.exports = [
 			call({ [a]: 5 });`,
 
 		output: `
-			call(( _obj = {}, _obj[a] = 5, _obj ));
-			var _obj;`
+			var _obj;
+
+			call(( _obj = {}, _obj[a] = 5, _obj ));`
 	},
 
 	{
@@ -160,7 +192,8 @@ module.exports = [
 	},
 
 	{
-		description: 'creates a computed method with a non-identifier expression (#78)',
+		description:
+			'creates a computed method with a non-identifier expression (#78)',
 
 		input: `
 			var obj = {
@@ -177,7 +210,8 @@ module.exports = [
 	},
 
 	{
-		description: 'does not require space before parens of computed method (#82)',
+		description:
+			'does not require space before parens of computed method (#82)',
 
 		input: `
 			var obj = {
@@ -194,7 +228,8 @@ module.exports = [
 	},
 
 	{
-		description: 'supports computed shorthand function with object spread in body (#135)',
+		description:
+			'supports computed shorthand function with object spread in body (#135)',
 
 		options: {
 			objectAssign: 'Object.assign'
@@ -217,15 +252,44 @@ module.exports = [
 	},
 
 	{
-		description: 'object literal with computed property within arrow expression (#126)',
+		description:
+			'object literal with computed property within arrow expression (#126)',
 
 		input: `
 			foo => bar({[x - y]: obj});
 		`,
 		output: `
-			(function(foo) { return bar(( _obj = {}, _obj[x - y] = obj, _obj ))
-				var _obj;; });
+			!function(foo) {
+				var _obj;
+
+				return bar(( _obj = {}, _obj[x - y] = obj, _obj ));
+			};
 		`
 	},
 
+	{
+		description: 'Supports nested computed properties (#51)',
+
+		input: `
+			(function () { return { [key]: { [key]: val } } })
+		`,
+		output: `
+			(function () {
+			var _obj, _obj$1;
+ return ( _obj$1 = {}, _obj$1[key] = ( _obj = {}, _obj[key] = val, _obj ), _obj$1 ) })
+		`
+	},
+
+	{
+		description: 'Puts helper variables in correct scope',
+
+		input: `
+			((x) => {var obj = 2; console.log([{[x]: 1}, obj]);})(3);
+		`,
+		output: `
+			(function (x) {
+			var _obj;
+var obj = 2; console.log([( _obj = {}, _obj[x] = 1, _obj ), obj]);})(3);
+		`
+	}
 ];