+++ test/samples/exponentiation-operator.js
@@ -6,19 +6,22 @@ module.exports = [
 	},
 
 	{
-		description: 'transpiles an exponentiation assignment to a simple reference',
+		description:
+			'transpiles an exponentiation assignment to a simple reference',
 		input: `x **= y`,
 		output: `x = Math.pow( x, y )`
 	},
 
 	{
-		description: 'transpiles an exponentiation assignment to a simple parenthesized reference',
+		description:
+			'transpiles an exponentiation assignment to a simple parenthesized reference',
 		input: `( x ) **= y`,
 		output: `( x ) = Math.pow( x, y )`
 	},
 
 	{
-		description: 'transpiles an exponentiation assignment to a rewritten simple reference',
+		description:
+			'transpiles an exponentiation assignment to a rewritten simple reference',
 
 		input: `
 			let x = 1;
@@ -38,7 +41,8 @@ module.exports = [
 	},
 
 	{
-		description: 'transpiles an exponentiation assignment to a simple member expression',
+		description:
+			'transpiles an exponentiation assignment to a simple member expression',
 
 		input: `
 			foo.bar **= y;`,
@@ -48,18 +52,22 @@ module.exports = [
 	},
 
 	{
-		description: 'transpiles an exponentiation assignment to a simple deep member expression',
+		description:
+			'transpiles an exponentiation assignment to a simple deep member expression',
 
 		input: `
 			foo.bar.baz **= y;`,
 
 		output: `
-			var object = foo.bar;
+			var object;
+
+			object = foo.bar;
 			object.baz = Math.pow( object.baz, y );`
 	},
 
 	{
-		description: 'transpiles an exponentiation assignment to a simple computed member expression',
+		description:
+			'transpiles an exponentiation assignment to a simple computed member expression',
 
 		input: `
 			foo[ bar ] **= y;`,
@@ -69,47 +77,57 @@ module.exports = [
 	},
 
 	{
-		description: 'transpiles an exponentiation assignment to a complex reference',
+		description:
+			'transpiles an exponentiation assignment to a complex reference',
 
 		input: `
 			foo[ bar() ] **= y;`,
 
 		output: `
-			var property = bar();
+			var property;
+
+			property = bar();
 			foo[property] = Math.pow( foo[property], y );`
 	},
 
 	{
-		description: 'transpiles an exponentiation assignment to a contrivedly complex reference',
+		description:
+			'transpiles an exponentiation assignment to a contrivedly complex reference',
 
 		input: `
 			foo[ bar() ][ baz() ] **= y;`,
 
 		output: `
-			var object = foo[ bar() ];
-			var property = baz();
+			var property, object;
+
+			object = foo[ bar() ];
+			property = baz();
 			object[property] = Math.pow( object[property], y );`
 	},
 
 	{
-		description: 'transpiles an exponentiation assignment to a contrivedly complex reference (that is not a top-level statement)',
+		description:
+			'transpiles an exponentiation assignment to a contrivedly complex reference (that is not a top-level statement)',
 
 		input: `
 			var baz = 1, lolwut = foo[ bar() ][ baz * 2 ] **= y;`,
 
 		output: `
-			var object, property;
+			var property, object;
+
 			var baz = 1, lolwut = ( object = foo[ bar() ], property = baz * 2, object[property] = Math.pow( object[property], y ) );`
 	},
 
 	{
-		description: 'transpiles an exponentiation assignment to a contrivedly complex reference with simple object (that is not a top-level statement)',
+		description:
+			'transpiles an exponentiation assignment to a contrivedly complex reference with simple object (that is not a top-level statement)',
 
 		input: `
 			var baz = 1, lolwut = foo[ bar() ] **= y;`,
 
 		output: `
 			var property;
+
 			var baz = 1, lolwut = ( property = bar(), foo[property] = Math.pow( foo[property], y ) );`
 	},
 
@@ -153,16 +171,18 @@ module.exports = [
 	},
 
 	{
-		description: 'handles assignment of exponentiation assignment to property with side effect',
+		description:
+			'handles assignment of exponentiation assignment to property with side effect',
 
 		input: `
 			x=a[bar()]**=2;
 		`,
 		output: `
 			var property;
+
 			x=( property = bar(), a[property]=Math.pow( a[property], 2 ) );
 		`
-	},
+	}
 
 	/* TODO: Test currently errors out with: TypeError: Cannot read property 'property' of null
 	{